+++ src/blueetl/extract/soma_report.py
@@ -6,6 +6,7 @@ import pandas as pd
 from bluepysnap import Simulation
 
 from blueetl.constants import CIRCUIT_ID, GID, NEURON_CLASS, SIMULATION_ID, TIME, VALUE, WINDOW
+from blueetl.core.utils import smart_concat
 from blueetl.extract.report import ReportExtractor
 
 L = logging.getLogger(__name__)
@@ -37,4 +38,4 @@ class SomaReport(ReportExtractor):
             df.rename(columns={0: VALUE}, inplace=True)
             df[WINDOW] = win.name
             df_list.append(df)
-        return df_list[0] if len(df_list) == 1 else pd.concat(df_list)
+        return smart_concat(df_list)